requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pydantic[email]
python-dotenv
python-jose[cryptography]
pyodbc
jose
python-dotenv
bcrypt